+++ apps/native-component-list/src/screens/ExpoApisScreen.tsx
@@ -49,6 +49,7 @@ const screens = [
   'FileSystem',
   'Font',
   'Errors',
+  'ExpoCoreModule',
   'ExpoModules',
   'Geocoding',
   'Haptics',