+++ packages/expo-modules-core/src/ts-declarations/global.ts
@@ -42,6 +42,11 @@ export interface ExpoGlobal {
    * or `null` if the view has not been registered.
    */
   getViewConfig(viewName: string): ViewConfig | null;
+
+  /**
+   * Reloads the app.
+   */
+  reloadAppAsync(reason: string): Promise<void>;
 }
 
 type ViewConfig = {